requirements.txt
selenium
jsbeautifier
python-mpv
PyInquirer
prompt_toolkit==1.0.14
tqdm
